requirements.txt
facebook-sdk==0.4.0
requests==2.3.0